refc : updated code version

+++ src/main/java/com/demo/application/controller/UDController.java
@@ -2,11 +2,13 @@ package com.demo.application.controller;
 
 import com.demo.application.dtos.AddUserDto;
 import com.demo.application.entity.UserEntity;
+import com.demo.application.exception.IdNotFoundException;
 import com.demo.application.service.UDService;
-//import io.swagger.v3.oas.annotations.responses.ApiResponse;
-//import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import com.demo.application.service.UDServiceImpl;
+import com.demo.application.util.Util;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,22 +18,25 @@ import java.util.List;
 @RestController
 //@Slf4j
 @RequestMapping("/api/ud")
+@AllArgsConstructor
 public class UDController{
 
-    @Autowired
     private UDService udService;
 
-    //@ApiResponses(value = {
-//        @ApiResponse(responseCode = "404",description = "there is no id")
-//})
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "404",description = "there is no id")
+})
     @PutMapping("/update/{id}")
     public ResponseEntity<String> updateUser(@RequestBody AddUserDto addUserDto,@PathVariable int id){
-        System.err.print("deleteUser");
+        if (!udService.existsUser(id))
+            throw Util.idNotFound();
         return new ResponseEntity<>(udService.updateUser(id,addUserDto).getBody().getFullName(), HttpStatus.OK);
     }
 
     @DeleteMapping("/delete")
     public ResponseEntity<String> deleteUser(@RequestParam int id){
+        if (!udService.existsUser(id))
+            throw Util.idNotFound();
         return new ResponseEntity<>(udService.deleteUser(id).getBody().getFullName(), HttpStatus.OK);
     }
